requirements.txt
Django==1.9.1
django-cors-headers==1.1.0
django-extensions==1.6.1
django-mathfilters==0.3.0
djangorestframework==3.3.2
functools32==3.2.3.post2
jsonschema==2.5.1
Pillow==3.1.1
six==1.10.0
